util: Fix printf format warning in piglit-vbo.cpp.
Fixes this GCC warning.
piglit-vbo.cpp: In constructor ‘vertex_attrib_description::vertex_attrib_description(GLuint, const char*)’:
piglit-vbo.cpp:221: warning: format ‘%lu’ expects type ‘long unsigned int’, but argument 2 has type ‘size_t’

+++ tests/util/piglit-vbo.cpp
@@ -218,7 +218,7 @@ vertex_attrib_description::vertex_attrib_description(GLuint prog,
 	}
 
 	if (this->count < 1 || this->count > 4) {
-		printf("Count must be between 1 and 4.  Got: %lu\n", count);
+		printf("Count must be between 1 and 4.  Got: %lu\n", (unsigned long) count);
 		piglit_report_result(PIGLIT_FAIL);
 	}
 }
